Move all Lambdas into a single project. This reduces code duplication.
