+++ event-processing/README.md
@@ -103,12 +103,12 @@ The event type we use for the event-processor lambda is an SQSEvent.
 Run functions locally and invoke them with the `sam local invoke` command.
 
 ```bash
-event-processor$ sam local invoke EventProcessorFunction --event events/event.json --env-vars invoke-vars/environment-vars.json --profile <dev acccount profile>
+event-processor$ sam local invoke KBVEventProcessorFunction --event invoke-events/event.json --env-vars invoke-vars/environment-vars.json --profile <dev acccount profile>
 ```
 You can also test against a Lambda deployed into the Dev environment using the AWS CLI:
 
 ```bash
-event-processor$ aws lambda invoke --function-name EventProcessorFunction --invocation-type Event --payload "<base64 encoded event json>" outfile.txt --profile <AWSProfileForTheTargetAccount>
+event-processor$ aws lambda invoke --function-name KBVEventProcessorFunction --invocation-type Event --payload "<base64 encoded event json>" outfile.txt --profile <AWSProfileForTheTargetAccount>
 ```
 
 ## Unit tests
