--- event-processing/event-processor/.eslintignore
@@ -1,2 +1,0 @@
-node_modules
-.aws-sam
