--- event-processing/event-processor/.npmignore
@@ -1,1 +1,0 @@
-tests/*
