+++ event-processing/lambda/.eslintignore
@@ -1,0 +1,2 @@
+node_modules
+.aws-sam
