+++ event-processing/lambda/enums/required-fields.enum.ts
