+++ event-processing/lambda/exceptions/hmac-exception.ts
