+++ event-processing/lambda/exceptions/validation-exception.ts
