+++ event-processing/lambda/models/log-details.interface.ts
