+++ event-processing/lambda/models/required-field-error.interface.ts
