+++ event-processing/lambda/models/required-fields-exception.interface.ts
