+++ event-processing/lambda/models/unknown-field-details.interface.ts
