+++ event-processing/lambda/models/validation-response.interface.ts
