+++ event-processing/lambda/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "event_processor",
+  "version": "1.0.0",
+  "description": "Event Processor for DI TxMA Audit",
+  "main": "app.js",
+  "repository": "https://github.com/alphagov/di-txma-audit",
+  "author": "GDS",
+  "license": "MIT",
+  "scripts": {
+    "unit": "jest",
+    "lint": "eslint '*.ts' --quiet --fix",
+    "compile": "tsc",
+    "test": "yarn run compile && yarn run unit"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-typescript"
+    ]
+  },
+  "devDependencies": {
+    "@babel/preset-typescript": "^7.16.7",
+    "@types/aws-lambda": "^8.10.92",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.13",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
+    "esbuild": "^0.14.14",
+    "eslint": "^8.8.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jest": "^27.5.0",
+    "prettier": "^2.5.1",
+    "ts-jest": "^27.1.3",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.5"
+  },
+  "dependencies": {
+    "aws-sdk": "^2.1136.0"
+  }
+}
