+++ event-processing/lambda/services/key-service.ts
