+++ event-processing/lambda/services/sns-service.ts
