+++ event-processing/lambda/services/validation-service.ts
