--- event-processing/re-ingest/.npmignore
@@ -1,1 +1,0 @@
-tests/*
