--- event-processing/re-ingest/app.ts
@@ -1,5 +1,0 @@
-import { SQSEvent } from 'aws-lambda';
-
-export const handler = async (event: SQSEvent): Promise<string> => {
-    return '';
-};
